ajuste de consultas e operações

+++ src/app/actions/infrastructure-component/destroy.ts
@@ -1,37 +1,30 @@
 'use server';
 
 import { localdatabase } from "@pedreiro-web/infrastructure/database/config";
+import { DockerControlPlane } from "@pedreiro-web/lib/docker";
 import { exec } from "child_process";
 
-export default async function DestroyInfrastructureComponent(prev: any, id: number) : Promise<{
+export default async function DestroyInfrastructureComponent(prev: any, id: number): Promise<{
     status: number
 } | undefined> {
+    const dockerControlPlane = new DockerControlPlane("/configuration/docker-compose.yml");
 
     const rows = localdatabase.prepare(`select * from infrastructure_component where id = ${id}`).all() as { service_key: string }[];
 
     localdatabase.exec(`insert into stream(operation, resource) values ('down', '${rows[0].service_key}')`);
 
-    const buildComponent = new Promise<string>((resolve, reject) => {
-        exec(`docker compose -f ./configuration/docker-compose.yml down ${rows[0].service_key}`,{ windowsHide: true }, (error, stdout, stderr) => {
-            if (error) {
-                console.log(error)
-                reject(`exec error: ${error}`);
-                return;
-            }
-            resolve("success");
-        })
-    });
-    
-    try {
-        await buildComponent;
-    } catch (ex) {
+    if (await dockerControlPlane.downService(rows[0].service_key)) {
+        return {
+            status: 200
+        }
+    } else {
         localdatabase.exec(`
-            DELETE FROM stream
-            WHERE resource = '${rows[0].service_key}' and operation = 'down';
-        `)
-    }
+                DELETE FROM stream
+                WHERE resource = '${rows[0].service_key}' and operation = 'down';
+            `)
 
-    return {
-        status: 200
+        return {
+            status: 400
+        }
     }
 }
